WIP: midway through MasterRotator/facade refactor
