+++ app/domain/rotation/master_rotator.rb
@@ -1,89 +1,6 @@
 module Rotation
   class MasterRotator
 
-    class Resource
-      attr_reader :id
-
-      def initialize(resource_id)
-        @id = resource_id
-      end
-
-      def account
-        @id.split(':')[0]
-      end
-
-      def kind
-        @id.split(':')[1]
-      end
-
-      def name
-        @id.split(':', 3)[2]
-      end
-
-      def renamed(name)
-        self.class.new("#{account}:#{kind}:#{name}")
-      end
-
-      #TODO add this
-      #
-      # prefix = variable_id.match(%r{(.*)/.*})[1]
-    end
-
-    class ScheduledRotation
-
-      RotatorNotFound = ::Util::ErrorClass.new(
-        "'{0}' is not an installed rotator"
-      )
-
-      attr_reader :resource, :ttl, :rotator_name
-
-      def initialize(resource_id:, ttl:, rotator_name:,
-                     avail_rotators:, secret_model:)
-        @resource = ::Rotation::MasterRotator::Resource.new(resource_id)
-        @ttl = ttl
-        @rotator_name = rotator_name
-        @avail_rotators = avail_rotators
-        @secret_model = secret_model
-        validate!
-      end
-
-      # This is a transformation pipeline, which we must express backwards in
-      # ruby:
-      # current_values | name_keys | new_values | resource_id_keys
-      #
-      def new_values
-        resource_id_keys(rotator.new_values(name_keys(current_values)))
-      end
-
-      private
-
-      def current_values
-        @secret_model.latest_resource_values(required_resources)
-        # p 'current_values', current_values
-      end
-
-      def rotator
-        @avail_rotators[rotator_name]
-      end
-
-      def required_resources
-        rotator.required_variables.map { |var| resource.renamed(var).id }
-      end
-
-      def validate!
-        raise RotatorNotFound, rotator_name unless rotator
-      end
-
-      def name_keys(vals_by_resource_id)
-        vals_by_resource_id.map do |resource_id, val|
-          [::Rotation::MasterRotator::Resource.new(resource_id).name, val]
-        end.to_h
-      end
-
-      def resource_id_keys(vals_by_name)
-        vals_by_name.map { |name, val| [resource.renamed(name).id, val] }.to_h
-      end
-    end
 
     # We inject both the rotation_model and the secret_model to ease unit
     # testing, and also because, even though the rotation query code was thrown
